Initial commit
Initial commit of project. Contains a rough sketch of the approach using
a tree of the possible connections in the abstract graph. This still
needs to be filled out and then another layer of "rendering" the graph
onto the board to see if it is viable must also be implemented. The
failure of both of these is the way we modify the tree. Guessing we'll
do some DFS (leaves are where both queues are empty) with a random
choice to decide which path to follow.