I have no idea
Pretty massive shift in thinking. This is just trying to get a good
model of how to layout memories for decisions as well as track the
current route. The meta now is to have the memories just be a series of
choices of which pieces were chosen and use the layer of the tree to
implicitly track constraints (so we can't put bad data in the memory
tree or we'll explode). Then we'll have a route_state that explicitly
tracks all that information. This route_state is what we'll use to
implement ascend, which just means undo previous decision (rise a level
in the memory tree). In this way we can just manipulate the route_state
and it will then persist these choices as memories.