Bugfix
Expand PathState and make it a member of Route. The reason we had to do
this is because otherwise we would trace down the tree, find a route,
return it. The next time we're invoked, we'd start at the root of the
tree. In this way, we'd only ever choose an initial vertex that we had
never chosen before. With PathState preserved, we can start at the
bottom. I think it should explore the valid paths now.